requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
text-unidecode==1.3
uritemplate==4.1.1
